Update week-01-murray-poem.md
Hi Jinhao! This poem looks great! There are a two more steps to this assignment that are missing - which is to find and display two images that go with your poem and to link to one youtube video. Add in those and you will be good to go. Review the Markdown how-to https://guides.github.com/features/mastering-markdown/#examples if you need a guide for how to add those things.

Let me know if you have questions!